OGR_GRASS: fix regression after switch to GetGDALDriverManager
OGR_GRASS driver was broken with 1f6a624e38a58cac80d754ae96131b5ee641bfda
as it was not working as it was, by registering with GetGDALDriverManager.

The unneeded OGRGRASSDriver class is removed and its use is replaced with
GDALDriver.

In related code 'bool', 'false' ,'true', 'nullptr' is used where appropriate.